Finished set-up of API requests
