+++ VirtueVerse/app/Http/Controllers/InsightsController.php
@@ -3,23 +3,39 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Services\InsightsApiService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 
 class InsightsController extends Controller
 {
-    public function getBookRecommendations(Request $request)
+    public function retrieveBookRecommendations(Request $request)
     {
+        $insightsService = new InsightsApiService();
 
+        $bookRecommendationsData = $insightsService->getBookRecommendations();
+
+        return response()->json($bookRecommendationsData);
     }
 
-    public function getExpectedCompletionTime(Request $request)
+    public function retrieveExpectedCompletionTime(Request $request)
     {
+        $insightsService = new InsightsApiService();
 
+        $completionTimedata = $insightsService->getExpectedCompletionTime();
+
+        return response()->json($completionTimedata);
     }
 
-    public function getPopularBooks(Request $request)
+    public function retrievePopularBooks(Request $request)
     {
-        
+        Log::info("Creating service");
+        $insightsService = new InsightsApiService();
+        Log::info("Finished creating service");
+
+        $popularBooksData = $insightsService->getPopularBooks();
+
+        return response()->json($popularBooksData);
     }
 }
