+++ VirtueVerse/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\InsightsController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -17,3 +18,7 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
+
+Route::post('/insights/retrieveBookRecommendations', [InsightsController::class, 'retrieveBookRecommendations'])->name('insights.retrieveBookRecommendations');
+Route::post('/insights/retrievePopularBooks', [InsightsController::class, 'retrievePopularBooks'])->name('insights.retrievePopularBooks');
+Route::post('/insights/retrieveExpectedCompletionTime', [InsightsController::class, 'retrieveExpectedCompletionTime'])->name('insights.retrieveExpectedCompletionTime');
